Fix issue 119 (SonarQube Administration General Settings Crashed) for…

+++ src/main/java/com/talanlabs/sonar/plugins/gitlab/GitLabPlugin.java
@@ -138,7 +138,7 @@ public class GitLabPlugin implements Plugin {
                         PropertyDefinition.builder(GITLAB_QUERY_MAX_RETRY).name("Query max retry").description("Max retry for wait finish analyse for publish mode").category(CATEGORY).subCategory(SUBCATEGORY)
                                 .type(PropertyType.INTEGER).defaultValue(String.valueOf(50)).index(28).build(),
                         PropertyDefinition.builder(GITLAB_QUERY_WAIT).name("Query waiting between retry").description("Max retry for wait finish analyse for publish mode (millisecond)").category(CATEGORY).subCategory(SUBCATEGORY)
-                                .type(PropertyType.LONG).defaultValue(String.valueOf(1000)).index(29).build(),
+                                .type(PropertyType.INTEGER).defaultValue(String.valueOf(1000)).index(29).build(),
                         PropertyDefinition.builder(GITLAB_QUALITY_GATE_FAIL_MODE).name("Quality Gate fail mode").description("Quality gate fail mode: error or warn")
                                 .category(CATEGORY).subCategory(SUBCATEGORY).type(PropertyType.SINGLE_SELECT_LIST)
                                 .options(QualityGateFailMode.WARN.getMeaning(), QualityGateFailMode.ERROR.getMeaning()).defaultValue(QualityGateFailMode.ERROR.getMeaning())
